+++ ChangeLog.md
@@ -9,6 +9,7 @@
 * [Feature] Ability to run jobs on timers while the application is in the foreground or app state changes (starting, resuming, or backgrounding)
 * [Enhancement] Increased discoverability via new AppDelegate & Android app/activity extension methods.   Simply add the Shiny namespace and type this.Shiny to see all of the points you should be attaching
 * [Enhancement][iOS][Android] Easier boilerplate setup
+* [Enhancement][Android] JobManager.RunTask will now use wakeful locks to run tasks if available
 * [Enhancement][Android] AndroidX support on android 10 targets - WorkManager replaces JobService under the hood
 * [Enhancement] Message bus name-only void events
 * [Enhancement] Connectivity now exposes cellular carrier